# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CTM API Docs",
    "version": "1.0.1",
    "description": "CTM all endpoints"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/buyNft/": {
      "post": {
        "tags": [
          "User",
          "Label"
        ],
        "description": "Endpoint for buying nfts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "You have just bought yourself a beautiful nft"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized - User is not logged in"
          },
          "403": {
            "description": "Forbidden - wrong role"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nftId": {
                    "type": "number",
                    "format": "int",
                    "example": 1
                  },
                  "amount": {
                    "type": "number",
                    "format": "int",
                    "example": 2
                  }
                },
                "required": [
                  "nftId",
                  "amount"
                ]
              }
            }
          }
        }
      }
    },
...
    "/buyCoins/buy": {
      "post": {
        "tags": [
          "User",
          "Label"
        ],
        "description": "Endpoint for buying the coins",
        "parameters": [
          {
            "name": "RequestBody",
            "in": "body",
            "required": true,
            "description": "Required request body to init transaction- signedTx is data(obtained from /initTransaction) which was signed by user wallet",
            "schema": {
              "type": "object",
              "properties": {
                "coinsAmount": {
                  "type": "number",
                  "example": 1
                },
                "signedTx": {
                  "type": "object",
                  "properties": {
                    "messageHash": {
                      "type": "string",
                      "example": "0xc454854490e717dded9ecd3f1cf78f12fbda542806c39be79120b69901415e92"
                    },
                    "v": {
                      "type": "string",
                      "example": "0x27125"
                    },
                    "r": {
                      "type": "string",
                      "example": "0x6503007bee9fce796d04f143bc5ce174e52094e12838ac3c78287ec518b1a93a"
                    },
                    "s": {
                      "type": "string",
                      "example": "0x5886e4ce7bcce1c6e4e242f73f6f7c81d9ecddec8254b4abda9922a9f020a288"
                    },
                    "rawTransaction": {
                      "type": "string",
                      "example": "0xf86e82029d84908a905182520894257b9eac215954863263bed86c65c4e642d00905867f544a44c0008083027125a06503007bee9fce796d04f143bc5ce174e52094e12838ac3c78287ec518b1a93aa05886e4ce7bcce1c6e4e242f73f6f7c81d9ecddec8254b4abda9922a9f020a288"
                    },
                    "transactionHash": {
                      "type": "string",
                      "example": "0x80a9b816d00572208d8c48f2d9f4a1a4e40181c0bcb0404628dfe9a042e01711"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response body at code 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "successfully bought 1 Moonlites"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "successfully bought 1 Moonlites"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized - User is not logged in"
          },
          "402": {
            "description": "Payment Required"
          },
          "403": {
            "description": "Forbidden - wrong role"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "coinsAmount": {
                    "example": "any"
                  },
                  "signedTx": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}